Update dependencies and organize folders
- remove functional widget package
- add provider architecture package
- add home-view-no-posts string to locals
- fix spelling errors
