+++ lib/core/managers/core_manager.dart
@@ -1,6 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:provider_start/core/managers/lifecycle_manager.dart';
 
+/// A Widget that will wrap all the managers into one widget
+///   - to add another manager just wrap the widget
+///     or add the new widget as a child of the current widget(s)
 class CoreManager extends StatelessWidget {
   final Widget child;
