+++ lib/core/models/user/user.dart
@@ -2,14 +2,14 @@ import 'dart:convert';
 
 import 'package:built_value/built_value.dart';
 import 'package:built_value/serializer.dart';
-import 'package:provider_start/core/serializers/serializers.dart';
+import 'package:provider_start/core/models/serializers.dart';
 
 part 'user.g.dart';
 
 /// An example user model that should be serialized.
 ///   - @nullable: means that its ok if the value is null
 ///   - @BuiltValueField: is the key that is in the JSON you
-///     recieve from an API
+///     receive from an API
 abstract class User implements Built<User, UserBuilder> {
   static Serializer<User> get serializer => _$userSerializer;
