+++ lib/core/repositories/posts_repository/posts_repository.dart
@@ -1,4 +1,4 @@
-import 'package:provider_start/core/serializers/post.dart';
+import 'package:provider_start/core/models/post/post.dart';
 
 abstract class PostsRepository {
   Future<List<Post>> fetchPosts();
