+++ lib/core/repositories/posts_repository/posts_repository_impl.dart
@@ -6,9 +6,9 @@ import 'package:provider_start/core/constant/api_routes.dart';
 import 'package:provider_start/core/constant/repository_exception_messages.dart';
 import 'package:provider_start/core/exceptions/network_exception.dart';
 import 'package:provider_start/core/exceptions/repository_exception.dart';
-import 'package:provider_start/core/hive_models/post_h.dart';
+import 'package:provider_start/core/models/post/post.dart';
+import 'package:provider_start/core/models/post/post_h.dart';
 import 'package:provider_start/core/repositories/posts_repository/posts_repository.dart';
-import 'package:provider_start/core/serializers/post.dart';
 import 'package:provider_start/core/services/connectivity/connectivity_service.dart';
 import 'package:provider_start/core/services/http/http_service_impl.dart';
 import 'package:provider_start/core/services/local_storage/local_storage_service.dart';
