+++ lib/core/repositories/users_repository/users_repository.dart
@@ -1,4 +1,4 @@
-import 'package:provider_start/core/serializers/user.dart';
+import 'package:provider_start/core/models/user/user.dart';
 
 abstract class UsersRepository {
   /// - throws `RepositoryException` if fetch fails
