+++ lib/core/repositories/users_repository/users_repository_impl.dart
@@ -5,9 +5,9 @@ import 'package:pedantic/pedantic.dart';
 import 'package:provider_start/core/constant/api_routes.dart';
 import 'package:provider_start/core/constant/repository_exception_messages.dart';
 import 'package:provider_start/core/exceptions/repository_exception.dart';
-import 'package:provider_start/core/hive_models/user_h.dart';
+import 'package:provider_start/core/models/user/user.dart';
+import 'package:provider_start/core/models/user/user_h.dart';
 import 'package:provider_start/core/repositories/users_repository/users_repository.dart';
-import 'package:provider_start/core/serializers/user.dart';
 import 'package:provider_start/core/services/connectivity/connectivity_service.dart';
 import 'package:provider_start/core/services/http/http_service_impl.dart';
 import 'package:provider_start/core/services/local_storage/local_storage_service.dart';
