+++ lib/core/view_models/base_view_model.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/widgets.dart';
 import 'package:provider_start/core/enums/view_state.dart';
 
-/// Class for viewmodels to extend from. Keeps track
+/// Class for view models to extend from. Keeps track
 /// of [ViewState] and exposes a function to change the
 /// state of the view.
-class BaseModel extends ChangeNotifier {
+class BaseViewModel extends ChangeNotifier {
   ViewState _state = ViewState.Idle;
   ViewState get state => _state;
