+++ lib/core/view_models/home_view_model.dart
@@ -1,11 +1,11 @@
 import 'package:provider_start/core/enums/view_state.dart';
 import 'package:provider_start/core/exceptions/repository_exception.dart';
+import 'package:provider_start/core/models/post/post.dart';
 import 'package:provider_start/core/repositories/posts_repository/posts_repository.dart';
-import 'package:provider_start/core/serializers/post.dart';
-import 'package:provider_start/core/ui_models/base_model.dart';
+import 'package:provider_start/core/view_models/base_view_model.dart';
 import 'package:provider_start/locator.dart';
 
-class HomeModel extends BaseModel {
+class HomeViewModel extends BaseViewModel {
   final _postsRepository = locator<PostsRepository>();
 
   List<Post> _posts = [];
