+++ lib/core/view_models/login_view_model.dart
@@ -5,10 +5,10 @@ import 'package:provider_start/core/exceptions/auth_exception.dart';
 import 'package:provider_start/core/mixins/validators.dart';
 import 'package:provider_start/core/services/auth/auth_service.dart';
 import 'package:provider_start/core/services/navigation/navigation_service.dart';
-import 'package:provider_start/core/ui_models/base_model.dart';
+import 'package:provider_start/core/view_models/base_view_model.dart';
 import 'package:provider_start/locator.dart';
 
-class LoginModel extends BaseModel with Validators {
+class LoginViewModel extends BaseViewModel with Validators {
   final _navigationService = locator<NavigationService>();
   final _authService = locator<AuthService>();
