+++ lib/ui/router.dart
@@ -2,15 +2,15 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_platform_widgets/flutter_platform_widgets.dart';
 import 'package:provider_start/core/constant/view_routes.dart';
-import 'package:provider_start/core/serializers/post.dart';
-import 'package:provider_start/ui/views/login/login_view.dart';
+import 'package:provider_start/core/models/post/post.dart';
+import 'package:provider_start/ui/views/login_view.dart';
 import 'package:provider_start/ui/views/main_view.dart';
-import 'package:provider_start/ui/views/post_details/post_details.dart';
-import 'package:provider_start/ui/views/splash/splash_view.dart';
+import 'package:provider_start/ui/views/post_details_view.dart';
+import 'package:provider_start/ui/views/splash_view.dart';
 
 /// Class that generates routes for the application
 ///   - Routes are generated by the list of routes in the app
-///   - Routes can also require paramaters. e.g. `PostDetailView(post: post)`
+///   - Routes can also require parameters. e.g. `PostDetailView(post: post)`
 class Router {
   static Route<dynamic> generateRoute(
     BuildContext context,
