+++ lib/ui/views/main_view.dart
@@ -1,18 +1,19 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_platform_widgets/flutter_platform_widgets.dart';
+import 'package:provider_architecture/provider_architecture.dart';
 import 'package:provider_start/core/localization/localization.dart';
-import 'package:provider_start/core/ui_models/views/main_model.dart';
-import 'package:provider_start/ui/views/base_view.dart';
-import 'package:provider_start/ui/views/home/home_view.dart';
-import 'package:provider_start/ui/views/settings/settings_view.dart';
+import 'package:provider_start/core/view_models/main_view_model.dart';
+import 'package:provider_start/ui/views/home_view.dart';
+import 'package:provider_start/ui/views/settings_view.dart';
 
 class MainView extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final local = AppLocalizations.of(context);
 
-    return BaseView<MainModel>(
+    return ViewModelProvider<MainViewModel>.withConsumer(
+      viewModel: MainViewModel(),
       builder: (context, model, child) => PlatformScaffold(
         body: IndexedStack(
           index: model.index,
