+++ lib/ui/views/post_details_view.dart
@@ -1,0 +1,65 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_platform_widgets/flutter_platform_widgets.dart';
+import 'package:provider_architecture/provider_architecture.dart';
+import 'package:provider_start/core/enums/view_state.dart';
+import 'package:provider_start/core/models/post/post.dart';
+import 'package:provider_start/core/view_models/post_details_view_model.dart';
+import 'package:provider_start/ui/shared/ui_helper.dart';
+
+class PostDetailsView extends StatelessWidget {
+  final Post post;
+
+  const PostDetailsView({Key key, @required this.post}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return ViewModelProvider<PostDetailsViewModel>.withConsumer(
+      viewModel: PostDetailsViewModel(),
+      onModelReady: (model) => model.init(post),
+      builder: (context, model, child) => PlatformScaffold(
+        appBar: PlatformAppBar(
+          title: Text(
+            post.title,
+            maxLines: 1,
+            overflow: TextOverflow.ellipsis,
+          ),
+        ),
+        body: Padding(
+          padding: const EdgeInsets.all(16.0),
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: <Widget>[
+              Text(post.description),
+              UIHelper.verticalSpaceLarge(),
+              _UserDetails(),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
+
+class _UserDetails extends ProviderWidget<PostDetailsViewModel> {
+  @override
+  Widget build(BuildContext context, PostDetailsViewModel model) {
+    if (model.state == ViewState.Busy) {
+      return Center(
+        child: PlatformCircularProgressIndicator(),
+      );
+    } else if (model.state == ViewState.Error) {
+      return Icon(Icons.error_outline);
+    }
+
+    return Column(
+      mainAxisSize: MainAxisSize.min,
+      children: <Widget>[
+        Text(model.user.username),
+        UIHelper.verticalSpaceMedium(),
+        Text(model.user.name),
+        UIHelper.verticalSpaceMedium(),
+        Text(model.user.website),
+      ],
+    );
+  }
+}
