+++ lib/ui/views/splash_view.dart
