+++ lib/ui/widgets/post_tile.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:provider_start/core/constant/view_routes.dart';
-import 'package:provider_start/core/serializers/post.dart';
+import 'package:provider_start/core/models/post/post.dart';
 
 class PostTile extends StatelessWidget {
   final Post post;
